MAGENTO2-188 Adjusted per pitbulk's PR suggestion
