+++ .php-cs.dist
@@ -22,13 +22,6 @@ $finder = Finder::create()
                 ->exclude('vendors')
                 ->in(__DIR__);
 
-/**
- * For PHP7.1< ignore this specific file
- */
-if (PHP_VERSION_ID < 70100) {
-    $finder->notPath('src/Controller/Plugin/Extendable/CsrfCompatible.php');
-}
-
 return Config::create()
              ->setRiskyAllowed(true)
              ->setUsingCache(true)
