+++ src/Controller/Plugin/Index.php
@@ -22,16 +22,72 @@
 
 namespace Shopgate\Base\Controller\Plugin;
 
+use Magento\Framework\App\Action\Action;
+use Magento\Framework\App\Action\Context;
+use Magento\Framework\App\Request\Http;
+use Magento\Framework\Module\ResourceInterface;
+use Shopgate\Base\Model\Forwarder;
+use Shopgate\Base\Model\Utility\Registry;
+
 /**
- * Conditional controller for backwards compatibility with Mage 2.3<
- * logic.
+ * Merchant API to Magento2 call router
  */
-if (interface_exists('\Magento\Framework\App\CsrfAwareActionInterface')) {
-    class Index extends Extendable\CsrfCompatible
-    {
+class Index extends Action
+{
+    /** @var Forwarder */
+    private $forwarder;
+    /** @var ResourceInterface */
+    private $versionInfo;
+    /** @var Registry */
+    private $registry;
+
+    /**
+     * @param Context           $context
+     * @param Forwarder         $forwarder
+     * @param ResourceInterface $versionInfo
+     * @param Registry          $registry
+     */
+    public function __construct(
+        Context $context,
+        Forwarder $forwarder,
+        ResourceInterface $versionInfo,
+        Registry $registry
+    ) {
+        $this->forwarder   = $forwarder;
+        $this->versionInfo = $versionInfo;
+        $this->registry    = $registry;
+
+        // CsrfAwareAction Magento2.3 compatibility
+        if (interface_exists('\Magento\Framework\App\CsrfAwareActionInterface')) {
+            $request = $context->getRequest();
+            if ($request instanceof Http && $request->isPost()) {
+                $request->setParam('isAjax', true);
+            }
+        }
+        parent::__construct($context);
     }
-} else {
-    class Index extends Extendable\Main
+
+    /**
+     * @throws \Exception
+     */
+    public function execute()
     {
+        define('SHOPGATE_PLUGIN_VERSION', $this->versionInfo->getDbVersion('Shopgate_Base'));
+        $this->registry->flagApi();
+        $this->registry->setAction($this->getRequest()->getParam('action'));
+
+        try {
+            $this->forwarder->handleRequest($this->getRequest()->getParams());
+        } catch (\ShopgateLibraryException $e) {
+            $response = new \ShopgatePluginApiResponseAppJson(
+                $this->getRequest()->getParam('trace_id')
+            );
+            $response->markError($e->getCode(), $e->getMessage());
+            $response->setData([]);
+            $response->send();
+        } catch (\Exception $e) {
+            throw new \Exception($e->getMessage());
+        }
     }
 }
+
